requirements.txt
appnope==0.1.4
asttokens==2.4.1
build==1.1.1
certifi>=2024.07.04
charset-normalizer==3.3.2
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
docutils==0.20.1
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.49.0
idna>=3.7
importlib-metadata==7.0.2
ipykernel==6.29.3
ipython==8.22.2
jaraco-classes==3.3.1
jedi==0.19.1
jupyter-client==8.6.0
jupyter-core==5.7.1
keyring==24.3.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.2.0
nest-asyncio==1.6.0
nh3==0.2.15
numpy==1.26.4
packaging==24.0
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow>=10.3.0
pip>=23.3
pkginfo==1.10.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.17.2
pyparsing==3.1.2
pyproject-hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gnupg==0.5.2
pytz==2024.1
pyzmq==25.1.2
readme-renderer==43.0
requests>=2.32.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
seaborn==0.13.2
setuptools>=70.0.0
six==1.16.0
sqlalchemy==2.0.28
sqlalchemy-utils==0.41.1
stack-data==0.6.3
tornado>=6.4.1
traitlets==5.14.1
twine==5.0.0
typing-extensions==4.10.0
tzdata==2024.1
urllib3>=2.2.2
wcwidth==0.2.13
wheel==0.41.0
zipp>=3.19.1
